requirements.txt
Django==1.9.7
appdirs==1.4.3
django-autoslug==1.9.3
django-bootstrap3==8.2.2
django-choices==1.5.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
django-debug-toolbar==1.7
django-extensions==1.7.8
django-filter==1.0.2
django-tenant-schemas==1.7.0
djangorestframework-jwt==1.10.0
djangorestframework-xml==1.3.0
djangorestframework==3.6.2
drfdocs==0.0.11
googlemaps==2.4.6
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
